requirements.txt
# Core dependencies
PyGithub==2.8.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# Scheduling
APScheduler==3.10.4

# Metrics and monitoring (optional, for future enhancements)
prometheus-client==0.19.0